+++ FHD/templates/admin-vue-dist/assets/js/package-C2L-ayWd.js
@@ -1,1 +1,0 @@
-const o="10.0.0",s={version:o};export{s as p};